requirements.txt
docopt
passlib
autobahn
pyasn1
aiopg
yoyo-migrations
sqlalchemy
requests
idna==2.5
aiohttp